requirements.txt
# RunPod API client dependencies
requests>=2.31.0

# AWS S3 integration
boto3>=1.28.0

# Image processing
Pillow>=10.0.0

# OpenAI GPT integration
openai>=1.0.0

# Environment variables
python-dotenv>=1.0.0

# Ngrok for tunneling
pyngrok>=7.0.0
